moved refresh function
