+++ x/icqoracle/keeper/icq.go
@@ -133,33 +133,6 @@ func OsmosisClPoolCallback(k Keeper, ctx sdk.Context, args []byte, query icqtype
 	return nil
 }
 
-// Main BeginBlocker function to refresh the price of a token (if applicable)
-func (k Keeper) RefreshTokenPrice(ctx sdk.Context, tokenPrice types.TokenPrice, updateIntervalSec uint64) error {
-	// Get last update time for this token
-	currentTime := ctx.BlockTime()
-	lastUpdate := tokenPrice.LastRequestTime
-	isNewToken := lastUpdate.IsZero()
-	updateIntervalPassed := currentTime.Sub(lastUpdate) >= time.Second*time.Duration(updateIntervalSec)
-
-	// If the update interval has not passed, don't update
-	if !isNewToken && !updateIntervalPassed {
-		return nil
-	}
-
-	// If never updated or update interval has passed, submit a new query for the price
-	// If a query was already in progress, it will be replaced with a new one that will
-	// have the same query ID
-	if err := k.SubmitOsmosisClPoolICQ(ctx, tokenPrice); err != nil {
-		return errorsmod.Wrapf(err,
-			"failed to submit Osmosis CL pool ICQ baseToken='%s' quoteToken='%s' poolId='%s'",
-			tokenPrice.BaseDenom,
-			tokenPrice.QuoteDenom,
-			tokenPrice.OsmosisPoolId)
-	}
-
-	return nil
-}
-
 func UnmarshalSpotPriceFromOsmosisClPool(tokenPrice types.TokenPrice, queryResponseBz []byte) (price math.LegacyDec, err error) {
 	var pool deps.OsmosisConcentratedLiquidityPool
 	if err := proto.Unmarshal(queryResponseBz, &pool); err != nil {
